+++ server/main.js
@@ -49,6 +49,10 @@ import "../imports/api/notifications/server/methods.js";
 
 import "../imports/api/search/server/methods.js";
 
+import "../imports/api/digests/digests.js";
+import "../imports/api/digests/server/methods.js";
+import "../imports/api/digests/server/jobs.js";
+
 import "./apm.js";
 
 if (Meteor.isServer) {