+++ test/main.js
@@ -11,6 +11,7 @@ import "/test/api/canvas/canvas.test";
 import "/test/api/healthReports/healthReports.test";
 import "/test/api/dashboards/dashboards.test";
 import "/test/api/search/search.test";
+import "/test/api/digests/digests.test";
 
 describe("latelier", function() {
   it("package.json has correct name", async function() {